fix(diiid): final-review polish — float64 timebase, stale-raw guard, doc/test fixes
- export.py: save raw_t_ms as float64. DIII-D time bases are absolute ms; at
  t≈4100 ms the float32 ULP (~0.49 µs) equals the 2 MS/s Mirnov sample period, so
  adjacent timestamps collapse into duplicates late in a shot. raw_x stays float32.
- spectrogram_view.py: a load without raw data now clears the cached raw trace and
  disables the toggle, so a previous shot's trace can't leak into the next shot's
  npz export.
- app/__main__.py: logo height 300 → 120 to match the CSS cap (kills dead space);
  reword the "daemon pool" comment (ThreadPoolExecutor threads are not daemon) and
  record the hung-fetch / shutdown trade-off.
- tests: float64 exact round-trip of a late-window timebase (export + GUI save);
  palette parity for viz.py _LINE_HEX/_MUTED_HEX; open the MDS_AVAILABLE gate in the
  CO2 routing test; refresh the stale Analyze-tab-scoping docstring.

+++ src/tokeye/export.py
@@ -136,7 +136,10 @@ def analysis_bundle(
 
     if raw is not None:
         raw_t_ms, raw_x = raw
-        bundle["raw_t_ms"] = np.asarray(raw_t_ms, dtype=np.float32)
+        # float64: DIII-D time bases are absolute ms; at t≈4100 ms the float32
+        # ULP (~0.49 µs) matches the 2 MS/s Mirnov sample period, collapsing
+        # adjacent timestamps into duplicates late in a shot. raw_x keeps float32.
+        bundle["raw_t_ms"] = np.asarray(raw_t_ms, dtype=np.float64)
         bundle["raw_x"] = np.asarray(raw_x, dtype=np.float32)
 
     return bundle
